requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
aiogram==3.21.0
pillow==11.3.0
python-slugify==8.0.4
itsdangerous==2.2.0
imageio-ffmpeg==0.6.0
